+++ internal/service/autoscalingplans/exports_test.go
@@ -1,0 +1,9 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package autoscalingplans
+
+// Exports for use in tests only.
+var (
+	FindScalingPlanByNameAndVersion = findScalingPlanByNameAndVersion
+)
